requirements.txt
PyQt5>=5.15.0
requests>=2.28.0
Pillow>=10.0.0
opencv-python>=4.8.0
